feat(frontend): redirect / to /catalog and promote stats bar to catalog pages
Closes F6 (master-tasks.md). ADR-011 specified a dedicated splash page with
hero copy, stats bar, and a top-10 preview table above a "View Full Catalog"
link; in practice this put an interstitial between users and the catalog —
the primary reason the site exists. The root route now redirects to
/catalog, and the aggregate stats block (formerly splash-only) renders at
the top of every catalog-flavor page.

- Extract StatsBar + private StatCard helper into components/catalog/StatsBar.tsx
  (reusable surface; F7 and F8 will extend this component rather than the pages)
- Collapse app/page.tsx to a three-line redirect to /catalog
- Render StatsBar at the top of app/catalog/page.tsx and app/search/page.tsx
- Add ADR-011 Amendment documenting the revised decision (ADR-030 policy:
  amend via separate file because the splash was in deployed infrastructure)

Also fixes a sorting defect on the former splash page: it pre-sorted the
full catalog by spectra_count desc, sliced to top 10, and handed those 10
rows to CatalogTable. Clicking a different column header re-sorted only
those 10 rows, not the full catalog. The redirect removes this code path
entirely — bugs_list_4_17_26.md item 2 closes as a side effect.

Closes F6
Closes bugs_list_4_17_26.md item 2
